token claims

+++ apps/user/user_repository.go
@@ -6,8 +6,8 @@ type userRepository interface {
 	Store(user User) (User, error)
 	Fetch() ([]User, error)
 	Update(user User) (User, error)
-	Delete(id int) error
-	FindById(id int) (User, error)
+	Delete(id uint) error
+	FindById(id uint) (User, error)
 	FindByEmail(email string) (User, error)
 }
 
@@ -42,7 +42,19 @@ func (r *repository) Fetch() ([]User, error) {
 }
 
 // Get User by Id
-func (r *repository) FindById(id int) (User, error) {
+func (r *repository) FindById(id uint) (User, error) {
+	// user := User{}
+	// if len(usersStorage) != int(id) {
+	// 	return user, errors.New("user doesn't exist")
+	// }
+
+	// for _, v := range usersStorage {
+	// 	if usersStorage[id-1] == v {
+	// 		user = v
+	// 		return user, errors.New("user doesn't exist")
+	// 	}
+	// 	user = usersStorage[id-1]
+	// }
 	return usersStorage[id-1], nil
 }
 
@@ -67,9 +79,9 @@ func (r *repository) Update(user User) (User, error) {
 }
 
 // Delete User
-func (r *repository) Delete(id int) error {
+func (r *repository) Delete(id uint) error {
 	usersStorage[id-1] = usersStorage[len(usersStorage)-1]
-	usersStorage[len(usersStorage)-1] = User{}
-	usersStorage = usersStorage[:len(usersStorage)-1]
+	usersStorage[uint(len(usersStorage))-1] = User{}
+	usersStorage = usersStorage[:uint(len(usersStorage))-1]
 	return nil
 }
